move state classes
